Remove hard dependency on graphql
