+++ graphql.d.ts
@@ -1,0 +1,2 @@
+
+export * from './dist/graphql'
