+++ package.json
@@ -44,7 +44,9 @@
     "cross-fetch": "^3.1.4",
     "eslint": "^8.14.0",
     "fs-extra": "^10.0.1",
+    "graphql": ">= 14",
     "graphql-tag": "^2.12.4",
+    "graphql-type-json": "^0.3.2",
     "jest": "^26.6.3",
     "nock": "^13.2.4",
     "prettier": "^2.2.1",
@@ -54,9 +56,11 @@
   },
   "dependencies": {
     "async-toolbox": "^0.7.3",
-    "graphql": ">= 14",
-    "graphql-type-json": "^0.3.2",
     "inflection": "^1.13.1",
     "yargs": "^16.2.0"
+  },
+  "peerDependencies": {
+    "graphql": ">= 14",
+    "graphql-type-json": ">= 0.3.2"
   }
 }
