+++ src/graphql/resolver-builder/content-type-resolver-builder.ts
@@ -1,8 +1,8 @@
 import type { Resolver, Resolvers } from '@apollo/client'
-import { generateUUID, idToName, isLinkContentTypeValidation, present, unionTypeDefName } from '../util'
-import { ContentfulDataSource } from '../dataSource'
-import { namespacedTypeName } from '../types'
-import { ContentType, ContentTypeField } from '../contentful/types'
+import { generateUUID, idToName, isLinkContentTypeValidation, present, unionTypeDefName } from '../../util'
+import { ContentfulDataSource } from '../../dataSource'
+import { namespacedTypeName } from '../../types'
+import { ContentType, ContentTypeField } from '../../contentful/types'
 
 export default class ContentTypeResolverBuilder {
