+++ src/graphql/resolver-builder/index.ts
@@ -1,8 +1,8 @@
 import type {Resolvers} from '@apollo/client'
 
-import { ContentfulDataSource } from '../dataSource'
-import { assetFieldResolver, namespacedTypeName } from '../types'
-import { ContentType } from '../contentful/types'
+import { ContentfulDataSource } from '../../dataSource'
+import { assetFieldResolver, namespacedTypeName } from '../../types'
+import { ContentType } from '../../contentful/types'
 import ContentTypeResolverBuilder from './content-type-resolver-builder'
 import QueryResolverBuilder from './query-resolver-builder'
