+++ src/graphql/schema-builder/index.ts
@@ -2,9 +2,9 @@ import { GraphQLFieldConfigMap, GraphQLObjectType, GraphQLSchema, GraphQLString 
 import inflection from 'inflection'
 
 import ContentTypeWriter from './content-type-writer'
-import { idToName } from '../util'
-import { ContentType } from '../contentful/types'
-import { namespace, Namespace, namespacedTypeName } from '../types'
+import { idToName } from '../../util'
+import { ContentType } from '../../contentful/types'
+import { namespace, Namespace, namespacedTypeName } from '../../types'
 
 export type SchemaBuilderOptions = {
   contentTypes: ContentType[]
